docs: Add a PDF documentation
Add an Rmd file to create a PDF documentation.
Add a folder containing the files used by this Rmd:
 - Description of the table columns, in a CSV file.
 - TeX files used to format the LaTeX file created during the process
   before turning it into a PDF file.
 - An image as an SVG file.
 - One reference table, in a CSV file.
Add to gitignore an intermediary PDF file created during the process to
be able to load the SVG file.
Add packages used by the Rmd document to DESCRIPTION.
Add the resulting PDF documentation in the inst/doc folder.